feat: move key/hash256 etc to its own file
